+++ StorePanel.cs
@@ -12,6 +12,23 @@ namespace TradersExtended
 {
     internal class StorePanel
     {
+        public class ItemToSell
+        {
+            public enum ItemType
+            {
+                Single,
+                Stack,
+                Combined
+            }
+
+            public ItemType itemType = ItemType.Single;
+            public ItemDrop.ItemData item;
+            public int stack;
+            public int price;
+            public int amount;
+            public int quality;
+        }
+
         private const float positionDelta = 35f;
 
         private static GameObject sellPanel;
@@ -29,11 +46,11 @@ namespace TradersExtended
 
         private static bool epicLootEnabled;
 
-        private static readonly List<GameObject> sellItemList = new List<GameObject>();
-        private static readonly List<ItemDrop.ItemData> tempItems = new List<ItemDrop.ItemData>();
-        private static readonly Dictionary<string, Dictionary<int, int>> tempItemsPrice = new Dictionary<string, Dictionary<int, int>>();
+        internal static readonly List<GameObject> sellItemList = new List<GameObject>();
+        internal static readonly List<ItemToSell> tempItems = new List<ItemToSell>();
+        internal static readonly Dictionary<string, Dictionary<int, int>> tempItemsPrice = new Dictionary<string, Dictionary<int, int>>();
 
-        private static ItemDrop.ItemData selectedItem;
+        private static ItemToSell selectedItem;
         private static int selectedItemIndex = -1;
 
         public static bool IsOpen()
@@ -59,9 +76,8 @@ namespace TradersExtended
         public static int GetSelectedItemIndex()
         {
             int result = -1;
-            List<ItemDrop.ItemData> availableItems = tempItems;
-            for (int i = 0; i < availableItems.Count; i++)
-                if (availableItems[i] == selectedItem)
+            for (int i = 0; i < tempItems.Count; i++)
+                if (tempItems[i] == selectedItem)
                     result = i;
 
             return result;
@@ -69,7 +85,7 @@ namespace TradersExtended
 
         public static void SelectItem(int index, bool center)
         {
-            if (sellItemList.Count == 0)
+            if (sellItemList.Count == 0 || index >= sellItemList.Count)
                 index = -1;
 
             center = center || index == 0;
@@ -85,27 +101,33 @@ namespace TradersExtended
 
         public static void SellSelectedItem(StoreGui __instance)
         {
-            if (ItemPrice(selectedItem) == 0)
-                return;
-
             ItemDrop m_coinPrefab = __instance.m_coinPrefab;
             if (m_coinPrefab == null)
             {
-                logger.LogWarning($"No m_coinPrefab is setted in StoreGui");
+                logger.LogWarning($"No m_coinPrefab set in StoreGui");
                 return;
             }
 
+            if (selectedItem == null)
+                return;
+
             selectedItemIndex = GetSelectedItemIndex();
 
-            GetSellItemPriceStack(selectedItem, out int itemPrice, out int itemStack);
+            if (selectedItem.itemType == ItemToSell.ItemType.Single)
+                Player.m_localPlayer.GetInventory().RemoveItem(selectedItem.item);
+            else if (selectedItem.itemType == ItemToSell.ItemType.Stack)
+                Player.m_localPlayer.GetInventory().RemoveItem(selectedItem.item.m_shared.m_name, selectedItem.stack, selectedItem.quality == 0 ? -1 : selectedItem.quality);
+            else if (selectedItem.itemType == ItemToSell.ItemType.Combined)
+                Player.m_localPlayer.GetInventory().RemoveItem(selectedItem.item.m_shared.m_name, selectedItem.amount, selectedItem.quality == 0 ? -1 : selectedItem.quality);
+            else
+                return;
 
-            Player.m_localPlayer.GetInventory().RemoveItem(selectedItem, itemStack);
-            Player.m_localPlayer.GetInventory().AddItem(m_coinPrefab.gameObject.name, itemPrice, m_coinPrefab.m_itemData.m_quality, m_coinPrefab.m_itemData.m_variant, 0L, "");
+            Player.m_localPlayer.GetInventory().AddItem(m_coinPrefab.gameObject.name, selectedItem.price, m_coinPrefab.m_itemData.m_quality, m_coinPrefab.m_itemData.m_variant, 0L, "");
             
-            string text = itemStack <= 1 ? selectedItem.m_shared.m_name : $"{selectedItem.m_stack}x{selectedItem.m_shared.m_name}";
+            string text = selectedItem.stack <= 1 ? selectedItem.item.m_shared.m_name : $"{selectedItem.stack}x{selectedItem.item.m_shared.m_name}";
 
             __instance.m_sellEffects.Create(__instance.transform.position, Quaternion.identity);
-            Player.m_localPlayer.Message(MessageHud.MessageType.TopLeft, Localization.instance.Localize("$msg_sold", text, itemPrice.ToString()), 0, selectedItem.m_shared.m_icons[0]);
+            Player.m_localPlayer.Message(MessageHud.MessageType.TopLeft, Localization.instance.Localize("$msg_sold", text, selectedItem.price.ToString()), 0, selectedItem.item.GetIcon());
             __instance.m_trader.OnSold();
             Gogan.LogEvent("Game", "SoldItem", text, 0L);
 
@@ -123,30 +145,129 @@ namespace TradersExtended
             if (!string.IsNullOrEmpty(item.notRequiredGlobalKey) && ZoneSystem.instance.GetGlobalKey(item.notRequiredGlobalKey))
                 return;
 
+            if (!TryGetPriceKey(item, out string key))
+                return;
+
+            if (!tempItemsPrice.ContainsKey(key))
+                tempItemsPrice[key] = new Dictionary<int, int>();
+
+            tempItemsPrice[key][item.stack] = item.price;
+        }
+
+        private static bool TryGetPriceKey(TradeableItem item, out string key)
+        {
+            key = "";
+
             GameObject prefab = ObjectDB.instance.GetItemPrefab(item.prefab);
             if (prefab == null)
-                return;
+                return false;
 
-            string name = prefab.GetComponent<ItemDrop>().m_itemData.m_shared.m_name;
+            ItemDrop.ItemData itemData = prefab.GetComponent<ItemDrop>().m_itemData;
+            if (itemData.m_shared.m_maxStackSize == 1 && item.stack != 1)
+                return false;
 
-            // While this data structure supports several price per stack current sell list implementation does not
-            // Maybe later
-            if (tempItemsPrice.ContainsKey(name))
-                return;
+            key = GetPriceKey(itemData, item.quality);
 
-            tempItemsPrice.Add(name, new Dictionary<int, int>());
+            return true;
+        }
 
-            if (!tempItemsPrice[name].ContainsKey(item.stack))
+        private static string GetPriceKey(ItemDrop.ItemData itemData, int quality)
+        {
+            return quality > 0 ? itemData.m_shared.m_name + "-" + quality : itemData.m_shared.m_name;
+        }
+
+        private static bool IgnoreItemForSell(ItemDrop.ItemData item)
+        {
+            if (item.m_shared.m_name == StoreGui.instance.m_coinPrefab.m_itemData.m_shared.m_name)
+                return true;
+            else if (!String.IsNullOrWhiteSpace(playerFilter.text) && Localization.instance.Localize(item.m_shared.m_name).ToLower().IndexOf(playerFilter.text.ToLower()) == -1)
+                return true;
+            else if (hideEquippedAndHotbarItems.Value && item.m_equipped) // Ignore currently equipped item
+                return true;
+            else if (hideEquippedAndHotbarItems.Value && item.m_gridPos.y == 0 && item.IsEquipable()) // Ignore equippable item from first row (hotbar)
+                return true;
+            else if (hideEquippedAndHotbarItems.Value && item.m_gridPos.y > Player.m_localPlayer.GetInventory().GetHeight()) // Ignore every additional (hidden) inventory row
+                return true;
+            else if (AzuExtendedPlayerInventory.API.GetQuickSlotsItems().Contains(item))
+                return true;
+
+            return false;
+        }
+
+        private static Dictionary<int, int> GetStackPrices(ItemDrop.ItemData item, out int quality)
+        {
+            quality = item.m_quality;
+            string key = GetPriceKey(item, quality);
+            if (tempItemsPrice.ContainsKey(key))
+                return tempItemsPrice[key];
+
+            quality = 0;
+            key = GetPriceKey(item, quality);
+            if (tempItemsPrice.ContainsKey(key))
+                return tempItemsPrice[key];
+
+            return null;
+        }
+
+        private static void AddToSellList(ItemDrop.ItemData item, int itemStack, int itemPrice, float priceFactor, int quality, ItemToSell.ItemType itemType)
+        {
+            int price = itemPrice;
+            if (itemType != ItemToSell.ItemType.Stack && qualityMultiplier.Value != 0 && quality == 0 && item.m_quality > 1)
+                price += (int)(qualityMultiplier.Value * price * (item.m_quality - 1));
+
+            price = Math.Max((int)(price * priceFactor), 1);
+
+            if (itemType == ItemToSell.ItemType.Single)
+                tempItems.Add(new ItemToSell()
+                {
+                    itemType = itemType,
+                    item = item,
+                    stack = itemStack,
+                    price = price,
+                    quality = quality
+                });
+            else if (itemType == ItemToSell.ItemType.Stack)
             {
-                if (item.stack == 1)
+                ItemToSell currentStack = tempItems.Find(tmpItem => tmpItem.itemType == itemType && 
+                                                                    tmpItem.stack == itemStack && 
+                                                                    tmpItem.item.m_shared.m_name == item.m_shared.m_name && 
+                                                                    (quality == 0 || tmpItem.item.m_quality == quality) && 
+                                                                    tmpItem.price == price);
+                if (currentStack != null)
+                    currentStack.amount += item.m_stack;
+                else
+                    tempItems.Add(new ItemToSell()
+                    {
+                        itemType = itemType,
+                        item = item,
+                        stack = itemStack,
+                        price = price,
+                        amount = item.m_stack,
+                        quality = quality
+                    });
+            }
+            else if (itemType == ItemToSell.ItemType.Combined)
+            {
+                ItemToSell currentStack = tempItems.Find(tmpItem => tmpItem.itemType == itemType && 
+                                                                    tmpItem.item.m_shared.m_name == item.m_shared.m_name && 
+                                                                    (quality == 0 || tmpItem.item.m_quality == quality) && 
+                                                                    tmpItem.stack == 1);
+                if (currentStack != null)
                 {
-                    Player.m_localPlayer.GetInventory().GetAllItems(name, tempItems);
-                    tempItemsPrice[name].Add(item.stack, item.price);
+                    currentStack.price += price;
+                    currentStack.amount += itemStack;
                 }
-                else if (Player.m_localPlayer.GetInventory().CountItems(name) >= item.stack)
+                else
                 {
-                    tempItems.Add(Player.m_localPlayer.GetInventory().GetItem(name));
-                    tempItemsPrice[name].Add(item.stack, item.price);
+                    tempItems.Add(new ItemToSell()
+                    {
+                        itemType = itemType,
+                        item = item,
+                        stack = 1,
+                        price = price,
+                        amount = itemStack,
+                        quality = quality
+                    });
                 }
             }
         }
@@ -167,56 +288,62 @@ namespace TradersExtended
             if (sellableItems.ContainsKey(CommonListKey(ItemsListType.Sell)))
                 sellableItems[CommonListKey(ItemsListType.Sell)].ForEach(item => AddItemToSellList(item));
 
-            bool filterSellList = !String.IsNullOrWhiteSpace(playerFilter.text);
-            for (int i = tempItems.Count - 1; i >= 0; i--)
+            float priceFactor = TraderCoins.GetPriceFactor(buyPrice: false);
+
+            foreach (ItemDrop.ItemData item in Player.m_localPlayer.GetInventory().GetAllItemsSorted())
             {
-                ItemDrop.ItemData tradeItem = tempItems[i];
-                if (tradeItem.m_shared.m_name == __instance.m_coinPrefab.m_itemData.m_shared.m_name)
-                    tempItems.RemoveAt(i);
-                else if (filterSellList && Localization.instance.Localize(tradeItem.m_shared.m_name).ToLower().IndexOf(playerFilter.text.ToLower()) == -1)
-                    tempItems.RemoveAt(i);
-                else if (hideEquippedAndHotbarItems.Value && tradeItem.m_equipped) // Ignore currently equipped item
-                    tempItems.RemoveAt(i);
-                else if (hideEquippedAndHotbarItems.Value && tradeItem.m_gridPos.y == 0 && tradeItem.IsEquipable()) // Ignore equippable item from first row (hotbar)
-                    tempItems.RemoveAt(i);
-                else if (hideEquippedAndHotbarItems.Value && tradeItem.m_gridPos.y > Player.m_localPlayer.GetInventory().GetHeight()) // Ignore every additional (hidden) inventory row
-                    tempItems.RemoveAt(i);
-                else if (AzuExtendedPlayerInventory.API.GetQuickSlotsItems().Contains(tradeItem))
-                    tempItems.RemoveAt(i);
+                if (IgnoreItemForSell(item))
+                    continue;
+
+                Dictionary<int, int> stackPrices = GetStackPrices(item, out int quality);
+                if (stackPrices == null)
+                    continue;
+
+                foreach (int stack in stackPrices.Keys.OrderBy(x => x))
+                    if (stack == 1)
+                        if (item.m_shared.m_maxStackSize == 1)
+                            AddToSellList(item, 1, stackPrices[stack], priceFactor, quality, ItemToSell.ItemType.Single);
+                        else
+                            AddToSellList(item, item.m_stack, stackPrices[stack] * item.m_stack, priceFactor, quality, ItemToSell.ItemType.Combined);
+                    else
+                        AddToSellList(item, stack, stackPrices[stack], priceFactor, quality, ItemToSell.ItemType.Stack);
             }
+
+            tempItems.RemoveAll(x => x.amount != 0 && x.amount < x.stack);
 
             float b = tempItems.Count * __instance.m_itemSpacing;
             b = Mathf.Max(__instance.m_itemlistBaseSize, b);
             listRoot.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, b);
             for (int i = 0; i < tempItems.Count; i++)
             {
-                ItemDrop.ItemData tradeItem = tempItems[i];
+                ItemToSell tradeItem = tempItems[i];
 
                 GameObject element = UnityEngine.Object.Instantiate(listElement, listRoot);
                 element.SetActive(value: true);
                 (element.transform as RectTransform).anchoredPosition = new Vector2(0f, (float)i * (0f - __instance.m_itemSpacing));
 
-                GetSellItemPriceStack(tradeItem, out int itemPrice, out int itemStack);
-
-                bool canSell = TraderCoins.CanSell(itemPrice);
+                bool canSell = TraderCoins.CanSell(tradeItem.price);
 
                 Image component = element.transform.Find("icon").GetComponent<Image>();
-                component.sprite = tradeItem.GetIcon();
+                component.sprite = tradeItem.item.GetIcon();
                 component.color = (canSell ? Color.white : new Color(1f, 0f, 1f, 0f));
 
-                string text = Localization.instance.Localize(tradeItem.m_shared.m_name);
+                string text = Localization.instance.Localize(tradeItem.item.m_shared.m_name);
 
-                if (qualityMultiplier.Value >= 0f && tradeItem.m_quality > 1)
-                    text += $" <color=#add8e6ff>({tradeItem.m_quality})</color>";
+                if (tradeItem.quality > 1)
+                    text += $" <color=#add8e6ff>({tradeItem.quality})</color>";
 
-                if (itemStack > 1)
-                    text += " x" + ItemStack(tradeItem);
+                if (tradeItem.stack > 1)
+                    text += " x" + tradeItem.stack;
+
+                if (tradeItem.amount > tradeItem.stack)
+                    text += $" <color=#c0c0c0ff>({tradeItem.amount})</color>";
 
                 TMP_Text component2 = element.transform.Find("name").GetComponent<TMP_Text>();
                 component2.SetText(text);
-                element.GetComponent<UITooltip>().Set(tradeItem.m_shared.m_name, tradeItem.GetTooltip(), __instance.m_tooltipAnchor);
+                element.GetComponent<UITooltip>().Set(tradeItem.item.m_shared.m_name, tradeItem.item.GetTooltip(), __instance.m_tooltipAnchor);
                 TMP_Text component3 = element.transform.Find("coin_bkg").Find("price").GetComponent<TMP_Text>();
-                component3.SetText(itemPrice.ToString());
+                component3.SetText(tradeItem.price.ToString());
                 if (!canSell)
                     component3.color = Color.grey;
 
@@ -234,50 +361,9 @@ namespace TradersExtended
             SelectItem(Mathf.Min(tempItems.Count - 1, selectedItemIndex), center: false);
         }
 
-        private static void GetSellItemPriceStack(ItemDrop.ItemData tradeItem, out int itemPrice, out int itemStack)
-        {
-            itemPrice = Math.Max((int)(ItemPrice(tradeItem) * TraderCoins.GetPriceFactor(buyPrice: false)), 1);
-            itemStack = ItemStack(tradeItem);
-        }
-
         public static void UpdateSellButton()
         {
-            sellButton.interactable = selectedItem != null;
-        }
-
-        private static int ItemPrice(ItemDrop.ItemData tradeItem)
-        {
-            if (tradeItem == null)
-                return 0;
-
-            int price = 0;
-            try
-            {
-                KeyValuePair<int, int> stackPrice = tempItemsPrice[tradeItem.m_shared.m_name].First();
-                price = (stackPrice.Key == 1) ? stackPrice.Value * tradeItem.m_stack : stackPrice.Value;
-                price += Math.Max((int)(qualityMultiplier.Value * price * (tradeItem.m_quality - 1)), 0);
-            }
-            catch (Exception ex)
-            {
-                logger.LogWarning($"Cannot find item {tradeItem.m_shared.m_name} price: {ex}");
-            }
-
-            return price;
-        }
-
-        private static int ItemStack(ItemDrop.ItemData tradeItem)
-        {
-            try
-            {
-                KeyValuePair<int, int> stackPrice = tempItemsPrice[tradeItem.m_shared.m_name].First();
-                return (stackPrice.Key == 1) ? tradeItem.m_stack : stackPrice.Key;
-            }
-            catch (Exception ex)
-            {
-                logger.LogWarning($"Cannot find item {tradeItem.m_shared.m_name} stack: {ex}");
-            }
-
-            return 1;
+            sellButton.interactable = selectedItem != null && TraderCoins.CanSell(selectedItem.price);
         }
 
         [HarmonyPatch(typeof(StoreGui), nameof(StoreGui.Awake))]
@@ -416,28 +502,6 @@ namespace TradersExtended
             }
         }
 
-        [HarmonyPatch(typeof(StoreGui), nameof(StoreGui.FillList))]
-        public static class StoreGui_FillList_FillSellableList
-        {
-            static void Postfix(StoreGui __instance, Trader ___m_trader, List<GameObject> ___m_itemList)
-            {
-                if (!modEnabled.Value) return;
-
-                if (epicLootEnabled)
-                {
-                    List<Trader.TradeItem> itemList = ___m_trader.GetAvailableItems();
-                    if (itemList.Count == ___m_itemList.Count)
-                        for (int i = 0; i < itemList.Count; i++)
-                        {
-                            Image component = ___m_itemList[i].transform.Find("icon").GetComponent<Image>();
-                            component.sprite = itemList[i].m_prefab.m_itemData.GetIcon();
-                        }
-                }
-
-                FillSellableList(__instance);
-            }
-        }
-
         [HarmonyPatch(typeof(Trader), nameof(Trader.GetAvailableItems))]
         public static class Trader_GetAvailableItems_FillBuyableItems
         {
@@ -473,60 +537,15 @@ namespace TradersExtended
                     {
                         __result[i] = JsonUtility.FromJson<Trader.TradeItem>(JsonUtility.ToJson(__result[i]));
                         __result[i].m_price = Math.Max((int)(__result[i].m_price * factor), 1);
-                        __result[i].m_stack = Math.Min(__result[i].m_stack, __result[i].m_prefab.m_itemData.m_shared.m_maxStackSize);
+                        TradeableItem.NormalizeStack(__result[i]);
                     }
             }
         }
 
         private static void AddAvailableItems(string listKey, List<Trader.TradeItem> __result)
         {
-            if (!tradeableItems.ContainsKey(listKey))
-                return;
-
-            foreach (TradeableItem item in tradeableItems[listKey])
-            {
-                if (!IsItemToSell(item, out ItemDrop prefab))
-                    continue;
-
-                if (__result.Exists(x => x.m_prefab == prefab))
-                {
-                    Trader.TradeItem itemTrader = __result.First(x => x.m_prefab == prefab);
-                    itemTrader.m_price = item.price;
-                    itemTrader.m_stack = item.stack;
-                    itemTrader.m_requiredGlobalKey = item.requiredGlobalKey;
-                }
-                else
-                {
-                    __result.Add(new Trader.TradeItem
-                    {
-                        m_prefab = prefab,
-                        m_price = item.price,
-                        m_stack = item.stack,
-                        m_requiredGlobalKey = item.requiredGlobalKey
-                    });
-                }
-            }
-        }
-
-        private static bool IsItemToSell(TradeableItem item, out ItemDrop itemDrop)
-        {
-            itemDrop = null;
-            if (!string.IsNullOrEmpty(item.requiredGlobalKey) && !ZoneSystem.instance.GetGlobalKey(item.requiredGlobalKey))
-                return false;
-
-            if (!string.IsNullOrEmpty(item.notRequiredGlobalKey) && ZoneSystem.instance.GetGlobalKey(item.notRequiredGlobalKey))
-                return false;
-
-            GameObject itemPrefab = ObjectDB.instance.GetItemPrefab(item.prefab);
-
-            if (itemPrefab == null)
-                return false;
-
-            itemDrop = itemPrefab.GetComponent<ItemDrop>();
-            if (itemDrop == null)
-                return false;
-
-            return !checkForDiscovery.Value || IgnoreItemDiscovery(item.prefab.ToLower()) || Player.m_localPlayer.IsMaterialKnown(itemDrop.m_itemData.m_shared.m_name);
+            if (tradeableItems.ContainsKey(listKey))
+                tradeableItems[listKey].DoIf(item => item.IsItemToSell(), item => __result.Add(item.ToTradeItem()));
         }
 
         [HarmonyPatch(typeof(StoreGui), nameof(StoreGui.UpdateSellButton))]
@@ -717,6 +736,79 @@ namespace TradersExtended
 
                 center = center || index == 0;
             }
-        }        
+        }
+
+        [HarmonyPatch(typeof(StoreGui), nameof(StoreGui.FillList))]
+        public static class StoreGui_FillList_FillSellableList
+        {
+            static bool Prefix(StoreGui __instance)
+            {
+                if (!modEnabled.Value)
+                    return true;
+
+                int playerCoins = __instance.GetPlayerCoins();
+                int num = __instance.GetSelectedItemIndex();
+                List<Trader.TradeItem> availableItems = __instance.m_trader.GetAvailableItems();
+                foreach (GameObject item in __instance.m_itemList)
+                {
+                    UnityEngine.Object.Destroy(item);
+                }
+
+                __instance.m_itemList.Clear();
+                float b = availableItems.Count * __instance.m_itemSpacing;
+                b = Mathf.Max(__instance.m_itemlistBaseSize, b);
+                __instance.m_listRoot.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, b);
+                for (int i = 0; i < availableItems.Count; i++)
+                {
+                    Trader.TradeItem tradeItem = availableItems[i];
+                    GameObject element = UnityEngine.Object.Instantiate(__instance.m_listElement, __instance.m_listRoot);
+                    element.SetActive(value: true);
+                    RectTransform rectTransform = element.transform as RectTransform;
+                    float num2 = (__instance.m_listRoot.rect.width - rectTransform.rect.width) / 2f;
+                    rectTransform.anchoredPosition = new Vector2(num2, i * (0f - __instance.m_itemSpacing) - num2);
+                    bool available = tradeItem.m_price <= playerCoins;
+                    Image component = element.transform.Find("icon").GetComponent<Image>();
+                    component.sprite = tradeItem.m_prefab.m_itemData.GetIcon();
+                    component.color = (available ? Color.white : new Color(1f, 0f, 1f, 0f));
+                    string text = Localization.instance.Localize(tradeItem.m_prefab.m_itemData.m_shared.m_name);
+
+                    TradeableItem.GetStackQualityFromStack(tradeItem.m_stack, out int stack, out int quality);
+                    if (quality > 1)
+                        text += $" <color=#add8e6ff>({quality})</color>";
+
+                    if (stack > 1)
+                        text += " x" + stack;
+
+                    TMP_Text component2 = element.transform.Find("name").GetComponent<TMP_Text>();
+                    component2.text = text;
+                    component2.color = available ? Color.white : Color.grey;
+
+                    string tooltip = ItemDrop.ItemData.GetTooltip(tradeItem.m_prefab.m_itemData, quality == 0 ? tradeItem.m_prefab.m_itemData.m_quality : quality, crafting: false, tradeItem.m_prefab.m_itemData.m_worldLevel);
+
+                    element.GetComponent<UITooltip>().Set(tradeItem.m_prefab.m_itemData.m_shared.m_name, tooltip, __instance.m_tooltipAnchor);
+                    TMP_Text component3 = Utils.FindChild(element.transform, "price").GetComponent<TMP_Text>();
+                    component3.text = tradeItem.m_price.ToString();
+                    if (!available)
+                        component3.color = Color.grey;
+
+                    element.GetComponent<Button>().onClick.AddListener(delegate
+                    {
+                        __instance.OnSelectedItem(element);
+                    });
+                    __instance.m_itemList.Add(element);
+                }
+
+                if (num < 0)
+                    num = 0;
+
+                __instance.SelectItem(Mathf.Min(__instance.m_itemList.Count - 1, num), center: false);
+
+                FillSellableList(__instance);
+
+                return false;
+            }
+        }
+
+
     }
 }